+++ src/pages/home/report/ReportActionCompose/AttachmentPickerWithMenuItems.tsx
@@ -18,7 +18,7 @@ import useTheme from '@hooks/useTheme';
 import useThemeStyles from '@hooks/useThemeStyles';
 import useWindowDimensions from '@hooks/useWindowDimensions';
 import * as Browser from '@libs/Browser';
-import getIconForAction from '@libs/focusComposerWithDelay/getIconForAction';
+import getIconForAction from '@libs/getIconForAction';
 import * as ReportUtils from '@libs/ReportUtils';
 import * as IOU from '@userActions/IOU';
 import * as Report from '@userActions/Report';
